+++ packages/schemas/src/v3/index.type.ts
@@ -369,6 +369,7 @@ export interface GerritConnectionConfig {
      */
     hidden?: boolean;
   };
+  revisions?: GitRevisions;
 }
 export interface BitbucketConnectionConfig {
   /**